month in long format

+++ src/Event.js
@@ -18,7 +18,7 @@ class Event extends Component {
     return <div className='event'>
       <h1 className='name'>{event.summary}</h1>
       <p>{new Date(event.start.dateTime).toLocaleDateString
-      ('en-US', { weekday: 'short', year: 'numeric', month: 'short', day: 'numeric', hour: 'numeric', minute: 'numeric', timeZone: 'utc', timeZoneName: 'short' })
+      ('en-US', { weekday: 'short', year: 'numeric', month: 'long', day: 'numeric', hour: 'numeric', minute: 'numeric', timeZone: 'utc', timeZoneName: 'short' })
       }
       </p>
       <p>
